requirements.txt
pyteal==0.18.1
beaker-pyteal==0.2.2
py-algorand-sdk==1.18.0
pytest==7.1.3
python-dotenv==0.21.0
pytealext==2.6.1
